+++ src/Test.java
@@ -34,8 +34,8 @@ public class Test {
 
         // load images
         BMPLoader bmp = new BMPLoader(indir, outdir);
-        // short[][] images = bmp.readAll();
-        short[][] images = bmp.readAll(NTRAIN);
+        // int[][] images = bmp.readAll();
+        int[][] images = bmp.readAll(NTRAIN);
         System.out.println("Elaborating images: " +
             images.length + "x" + bmp.height + "x" + bmp.width);
 
@@ -47,10 +47,10 @@ public class Test {
         }
         bmp.saveAll(svq.getData(), "centr");
 
-        short[][] selected = new short[NVALID][];
-        short[][] codes    = new short[NVALID][];
-        short[][] reconstr = new short[NVALID][];
-        short[][] errors   = new short[NVALID][];
+        int[][] selected = new int[NVALID][];
+        int[][] codes    = new int[NVALID][];
+        int[][] reconstr = new int[NVALID][];
+        int[][] errors   = new int[NVALID][];
 
         // NVALID "pseudorandom" images, from training range
         for (int i=0; i<NVALID; i++) {
